+++ src/test/java/org/opensearch/knn/index/codec/KNNCodecTestCase.java
@@ -212,7 +212,8 @@ public class KNNCodecTestCase extends KNNTestCase {
             ZonedDateTime.now(ZoneOffset.UTC).toString(),
             "",
             "",
-            ""
+            "",
+            MethodComponentContext.EMPTY
         );
 
         Model mockModel = new Model(modelMetadata1, modelBlob, modelId);